LIKA-229: Refactor package_generator into utils
