+++ ckanext/ckanext-apicatalog_ui/ckanext/apicatalog_ui/utils.py
@@ -1,0 +1,20 @@
+import ckan.plugins.toolkit as toolkit
+import itertools
+
+
+def package_generator(context, query='*:*', page_size=1000, dataset_type='dataset'):
+    package_search = toolkit.get_action('package_search')
+
+    # Loop through all items. Each page has {page_size} items.
+    # Stop iteration when all items have been looped.
+    for index in itertools.count(start=0, step=page_size):
+        data_dict = {'include_private': True, 'rows': page_size, 'q': query, 'start': index,
+                     'fq': '+dataset_type:' + dataset_type}
+        data = package_search(context, data_dict)
+        packages = data.get('results', [])
+        for package in packages:
+            yield package
+
+        # Stop iteration all query results have been looped through
+        if data["count"] < (index + page_size):
+            return
